// backend/enums/CamposBD.ts
enum CamposBD {
    A_ID_ALL_USER = "all",
    
    PERSON_ID_PERSON = "person.id_person",
    PERSON_FIRST_NAME = "person.first_name",
    PERSON_LAST_NAME = "person.last_name",
    PERSON_EMAIL = "person.email",
    PERSON_PASSWD = "person.passwd",
    PERSON_TOKEN_SESSION = "person.token_session",

    A_ID_PERSON = "id_person",
    A_FIRST_NAME = "first_name",
    A_LAST_NAME = "last_name",
    A_EMAIL = "email",
    A_PASSWD = "passwd",
    A_TOKEN_SESSION = "token_session",

    ADMINS_ID_ADMIN = "admins.id_admin",
    ADMINS_LAST_LOGIN = "admins.last_login",
    ADMINS_ACCESS_LEVEL = "admins.access_level",

    A_ID_ADMIN = "id_admin",
    A_LAST_LOGIN = "last_login",
    A_ACCESS_LEVEL = "access_level",

    CUSTOMER_ID_CUSTOMER = "customer.id_customer",
    CUSTOMER_PHONE = "customer.phone",
    CUSTOMER_DATE_BIRTH = "customer.date_birth",
    CUSTOMER_DOCUMENT_TYPE = "customer.document_type",
    CUSTOMER_DOCUMENT_NUM = "customer.document_num",
    CUSTOMER_IS_ALLOWED = "customer.is_allowed",

    A_ID_CUSTOMER = "id_customer",
    A_PHONE = "phone",
    A_DATE_BIRTH = "date_birth",
    A_DOCUMENT_TYPE = "document_type",
    A_DOCUMENT_NUM = "document_num",
    A_IS_ALLOWED = "is_allowed",

    LESSOR_ID_LESSOR = "lessor.id_lessor",
    LESSOR_REGISTRARION_DATE = "lessor.registration_date",

    A_ID_LESSOR = "id_lessor",
    A_REGISTRARION_DATE = "registration_date",
    
    ADDRESSES_ID_ADDRESS = "addresses.id_address",
    ADDRESSES_ADDRESS_NAME = "addresses.address_name",
    ADDRESSES_ZIP_CODE = "addresses.zip_code",
    ADDRESSES_CITY = "addresses.city",
    ADDRESSES_COUNTY = "addresses.county",
    ADDRESSES_STATE_NAME = "addresses.state_name",
    ADDRESSES_COUNTRY = "addresses.country",

    A_ID_ADDRESS = "id_address",
    A_ADDRESS_NAME = "address_name",
    A_ZIP_CODE = "zip_code",
    A_CITY = "city",
    A_COUNTY = "county",
    A_STATE_NAME = "state_name",
    A_COUNTRY = "country",

    SPORTFIELD_ID_SPORTFIELD = "sportfield.id_sportfield",
    SPORTFIELD_SPORTFIELD_NAME = "sportfield.sportfield_name",
    SPORTFIELD_CAPACITY = "sportfield.capacity",

    A_ID_SPORTFIELD = "id_sportfield",
    A_SPORTFIELD_NAME = "sportfield_name",
    A_CAPACITY = "capacity",

    BOOKING_ID_BOOKING = "booking.id_booking",
    BOOKING_BOOKING_DATE = "booking.booking_date",
    BOOKING_START_TIME = "booking.start_date",
    BOOKING_END_TIME = "booking.end_date",

    A_ID_BOOKING = "id_booking",
    A_BOOKING_DATE = "booking_date",
    A_START_TIME = "start_date",
    A_END_TIME = "end_date"
}

export default CamposBD